resolve clippy issues renamed raster/raster to raster/base_raster
